Make eventual migration to OpenMRS 1.11 easier.
I was playing around with updating to OpenMRS 1.11 a few weeks ago. Whilst we can't upgrade
in our current state (we depend on an OpenMRS 1.10 DB snapshot) we _can_ fix some of the smaller
blockers.

This change fixes the test data to work in 1.11, which introduces DB constraint checks for tests,
and adds comments cross-referencing when identical versions have been defined in more than one
place.