Add the enterer's UUID to observations sent from the server.
This allows clients to properly attribute observations to a specific user.

Note that we're especially carefully not to crash sync if an observation was recorded by someone
that doesn't exist in a way that's compatible with the Buendia model of a 'User'.

Each Buendia user has the following three objects in OpenMRS:
- A Person, which contains information such as the person's name
- A User, which contains login credentials (though Buendia doesn't really populate these properly)
- A Provider, which is the OpenMRS representation of a person who administers orders.

When users are synced from the server to the client, the Provider ID gets sent down at the moment,
so we have to do a bit of work to convert providers into users and back throughout.