Change diffdeb so as to ignore file timestamps.
This is necessary now that we're building packages in CI. CircleCI
regularly makes a fresh checkout of the code base, with the result that
file timestamps are different from one CI run to the next, because git
doesn't set them.

This change ensures that file timestamps are never a reason in and of
themselves to retain an otherwise unchanged Debian package.