Make it possible to trigger_archive_update against a specific CircleCI job.
This is necessary to enable manual archive update against a tagged
release build.

This code is needed mostly because the CircleCI config for build #185
(https://circleci.com/gh/projectbuendia/buendia/185) which built tag
v0.10.0 did not have the correct configuration to trigger the
apt-archive job to rebuild the apt repo in projectbuendia.github.io.