Rename "mock" to "dummy" because that's what this package really is.
<ping> Terminology nit: I'd like to be precise about the words we use
for test doubles. A "mock" is a stand-in for an object whose purpose is
to verify that an expected sequence of methods is called on the object.
A "fake" is a stand-in that replaces something complex with simplified
or minimal behaviour. A "dummy" is a stand-in that doesn't attempt to
simulate the behaviour of any existing program component.

https://github.com/projectbuendia/buendia/pull/190#discussion_r318321665