Clean up config.d/10-networking.
In particular, don't just assume that we always have a Wi-Fi interface.

If we *do*, then treat it as the main interface for talking to clients,
and treat the ethernet as a possible backhaul connection.

If we *don't*, then treat the ethernet connection as the main interface
for talking to clients.

Also, clean up some other stuff that wasn't being used, or the logic
didn't quite make sense.