wpa_cli doesn't work on the NUC; use iwconfig instead.
Currently buendia-wifi-watchdog uses wpa_cli to determine if a wifi
interface is associated to an access point. However on the NUC this
command always returns "disconnected", causing the watchdog to restart
the connection once every minute.

Instead, let's use iwconfig and ping to see if we're on a wifi network
and can talk to anyone.